Great overhaul of the whole architecture
The previous model was too impractical, too many tiny steps in
isolated files. Now There are fewer and longer files, but still the
various main stages of the installation process are rather well
separated.

Added a gitignored folder dev/ to contain the experiences I make
(trying to make everything work) and their results, before they're
being published in the rest of the project.